Execute all GUI related tasks in the event dispatch thread
In the prior versions all GUI manipulations were executed in the main
thread in contrast to the event handlers, which were executed in the
event dispatch thread. It could cause nasty race condition issues
worsening UX. In particular, CancelDialog sometimes failed to
disappear and didn't return the control back to the main frame.