Fix graph flickering
In prior versions the graph rendering was event driven. It means the
graph was rendered during the initialization and then re-rendered
by the mouse even handler. That inconsistent frame pacing resulted
in disturbing flickering.

The solution is to update the graph at a fixed rate. To implement this
the Graph constructor creates a thread which calls the repaint method
every 16ms ( 60fpf ). The information about how the graph is drawn
has to be prepared beforehand to be displayed during the next
scheduled repainting. But that solution comes at a cost of
unnoticeable IO latency.