Add a mechanism of removing settings
When a value hadn't yet been set SettingManipulator.getValue would
have returned an empty optional. However, if the value had been set to
an empty string, getValue would still have treated it as if the
value had been set. Setting the value to an empty string to remove
it would streamline the way of deleting it compare to adding
removeValue method. That means any empty string values are now
converted to an empty optional upon getValue invocation.