Fix SettingLoader config parsing
The config looks like a set of ordered pairs separated with new
lines. The ordered pair is a key and a value separated by one or more
whitespaces.

http_url example.org
protocol http

That is 2 line example of such config. In prior version invocation of
SettingLoader.getValue("url") would've returned "example.org" even so
that parameter is clearly not set. It was due the lousy use of regex,
which was targeting substrings of parameters.