Fix a possible memory leak in the unit tests
In essence, HTTP200Response is a web server running in a separate
thread. Its initialization allocates a lot of resources as well as
borrows a port. After being started it waits for a client connection.
But if no client connection is established the thread was basically
uninterruptible. As a result it wasn't gc'ed as well as didn't allow
other instances to use the same port. There was also no closing of the
established connections.

To fix this a timeout was set, so connection waiting is constantly
being interrupted. But then it starts waiting again if the condition
of the while loop is still true. To assure that the web server was
stopped HTTP200Response.join() needs to be invoked after
HTTP200Response.interrupt().