Fix IO socket lock in FloodWebServer
After being started, FloodWebServer creates a server sockets and waits
for an incoming connection. That waiting blocked the thread which was
why FloodWebServer.interrupt() didn't actually interrupted the thread.
A timeout prevents the infinite locking from happening. The invocation
of Thread.join() was added to FloodWebServer.interrupt() to make sure
the timeout reaches its limit making the thread to escape the while
loop.